Refactor: 도메인별 패키지 구조로 리팩토링
