+++ src/main/java/org/sopt/article/domain/Tag.java
@@ -1,0 +1,8 @@
+package org.sopt.article.domain;
+
+public enum Tag {
+    CS,
+    DB,
+    SPRING,
+    ETC
+}
