--- src/main/java/org/sopt/domain/Gender.java
@@ -1,7 +1,0 @@
-package org.sopt.domain;
-
-public enum Gender {
-    MALE,
-    FEMALE,
-    OTHER
-}
