--- src/main/java/org/sopt/domain/Tag.java
@@ -1,8 +1,0 @@
-package org.sopt.domain;
-
-public enum Tag {
-    CS,
-    DB,
-    SPRING,
-    ETC
-}
