--- src/main/java/org/sopt/dto/MemberDeleteResponse.java
@@ -1,8 +1,0 @@
-package org.sopt.dto;
-
-public record MemberDeleteResponse(Long deletedUserId
-) {
-    public static MemberDeleteResponse of(Long userId) {
-        return new MemberDeleteResponse(userId);
-    }
-}
