--- src/main/java/org/sopt/dto/MemberResponse.java
@@ -1,24 +1,0 @@
-package org.sopt.dto;
-
-import org.sopt.domain.Gender;
-import org.sopt.domain.Member;
-
-import java.time.LocalDate;
-
-public record MemberResponse(
-        Long userId,
-        String name,
-        LocalDate birthdate,
-        String email,
-        Gender gender
-) {
-    public static MemberResponse from(Member member) {
-        return new MemberResponse(
-                member.getId(),
-                member.getName(),
-                member.getBirthdate(),
-                member.getEmail(),
-                member.getGender()
-        );
-    }
-}
