--- src/main/java/org/sopt/exception/DataStorageException.java
@@ -1,7 +1,0 @@
-package org.sopt.exception;
-
-public class DataStorageException extends RuntimeException {
-    public DataStorageException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
