--- src/main/java/org/sopt/exception/DuplicateMemberException.java
@@ -1,7 +1,0 @@
-package org.sopt.exception;
-
-public class DuplicateMemberException extends IllegalStateException {
-    public DuplicateMemberException(String message) {
-        super(message);
-    }
-}
