--- src/main/java/org/sopt/exception/EntityNotFoundException.java
@@ -1,7 +1,0 @@
-package org.sopt.exception;
-
-public class EntityNotFoundException extends RuntimeException {
-    public EntityNotFoundException(String message) {
-        super(message);
-    }
-}
