--- src/main/java/org/sopt/exception/InvalidCsvFormatException.java
@@ -1,11 +1,0 @@
-package org.sopt.exception;
-
-public class InvalidCsvFormatException extends RuntimeException {
-    public InvalidCsvFormatException(String message) {
-        super(message);
-    }
-
-    public InvalidCsvFormatException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
