--- src/main/java/org/sopt/exception/InvalidEmailFormatException.java
@@ -1,7 +1,0 @@
-package org.sopt.exception;
-
-public class InvalidEmailFormatException extends IllegalArgumentException {
-    public InvalidEmailFormatException(String message) {
-        super(message);
-    }
-}
