--- src/main/java/org/sopt/exception/MemberAgeException.java
@@ -1,7 +1,0 @@
-package org.sopt.exception;
-
-public class MemberAgeException extends IllegalStateException {
-    public MemberAgeException(String message) {
-        super(message);
-    }
-}
