+++ src/main/java/org/sopt/global/component/IdGenerator.java
@@ -1,0 +1,6 @@
+package org.sopt.global.component;
+
+public interface IdGenerator {
+    long generate();
+    void initialize(long initialValue);
+}
