+++ src/main/java/org/sopt/member/dto/MemberDeleteResponse.java
@@ -1,0 +1,8 @@
+package org.sopt.member.dto;
+
+public record MemberDeleteResponse(Long deletedUserId
+) {
+    public static MemberDeleteResponse of(Long userId) {
+        return new MemberDeleteResponse(userId);
+    }
+}
