+++ src/main/java/org/sopt/member/dto/MemberListResponse.java
@@ -1,0 +1,15 @@
+package org.sopt.member.dto;
+
+import org.sopt.member.domain.Member;
+
+import java.util.List;
+
+public record MemberListResponse(List<MemberResponse> members
+) {
+    public static MemberListResponse from(List<Member> memberList) {
+        List<MemberResponse> memberDtos = memberList.stream()
+                .map(MemberResponse::from)
+                .toList();
+        return new MemberListResponse(memberDtos);
+    }
+}
