+++ src/main/java/org/sopt/member/exception/DataStorageException.java
@@ -1,0 +1,7 @@
+package org.sopt.member.exception;
+
+public class DataStorageException extends RuntimeException {
+    public DataStorageException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
