+++ src/main/java/org/sopt/member/exception/DuplicateMemberException.java
@@ -1,0 +1,7 @@
+package org.sopt.member.exception;
+
+public class DuplicateMemberException extends IllegalStateException {
+    public DuplicateMemberException(String message) {
+        super(message);
+    }
+}
