+++ src/main/java/org/sopt/member/exception/EntityNotFoundException.java
@@ -1,0 +1,7 @@
+package org.sopt.member.exception;
+
+public class EntityNotFoundException extends RuntimeException {
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
