+++ src/main/java/org/sopt/member/exception/InvalidCsvFormatException.java
@@ -1,0 +1,11 @@
+package org.sopt.member.exception;
+
+public class InvalidCsvFormatException extends RuntimeException {
+    public InvalidCsvFormatException(String message) {
+        super(message);
+    }
+
+    public InvalidCsvFormatException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
