+++ src/main/java/org/sopt/member/exception/InvalidEmailFormatException.java
@@ -1,0 +1,7 @@
+package org.sopt.member.exception;
+
+public class InvalidEmailFormatException extends IllegalArgumentException {
+    public InvalidEmailFormatException(String message) {
+        super(message);
+    }
+}
