+++ src/main/java/org/sopt/member/exception/MemberAgeException.java
@@ -1,0 +1,7 @@
+package org.sopt.member.exception;
+
+public class MemberAgeException extends IllegalStateException {
+    public MemberAgeException(String message) {
+        super(message);
+    }
+}
