--- src/main/java/org/sopt/repository/MemberRepository.java
@@ -1,10 +1,0 @@
-package org.sopt.repository;
-
-import org.sopt.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByEmail(String email);
-}
