--- src/main/java/org/sopt/service/component/IdGenerator.java
@@ -1,6 +1,0 @@
-package org.sopt.service.component;
-
-public interface IdGenerator {
-    long generate();
-    void initialize(long initialValue);
-}
